Move queries to components folders
