+++ src/components/Landing/components/LandingTop.jsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { Link } from "react-router-dom";
 import { client } from "../../../index"
-import voyagesQuery from "../../../queries/voyagesQuery"
+import voyagesQuery from "../../VoyagePortal/graphql/voyagesQuery"
+
 export default props => (
   <div className="landing-top" >
     <div className="tagline-box" >
