--- src/components/VoyagePortal/graphql/query.js
@@ -1,19 +1,0 @@
-import gql from "graphql-tag";
-export const get_voyages = gql`
-  query getVoyages {
-    cohorts {
-        id
-        title
-        start_date
-        end_date
-        status
-        members {
-          id
-          status
-          user {
-            id
-          }
-        }
-    }
-  }
-`
